Recreate old posts
Imports all existing posts from my old Jekyll site.
The latest post that was copied is from 2019-02-18.
There are two posts that exist on holgerfrohloff.de that didn't exist in
the repo:

* 2019-02-14 List of programming principles
* 2019-02-15 2018 state of code review study

Those will have to be copied over manually. Afterwards, I can probably
take the existing posts from Joplin and recreate these as markdown files
in the repo.

+++ importer.rb
@@ -1,23 +1,29 @@
 #!/usr/bin/env ruby
-
+require "fileutils"
 source_dir = "/Users/holger/projects/site/_posts"
 target_dir = "/Users/holger/projects/gatsby-site/content/posts"
 
 date_regex = /\d{4}-\d{2}-\d{2}+/
-title_regex = /(?:\d-)*(([\D+-])+)(\.markdown)/
+title_regex = /(?:\d-)*(([\D+-])+)0?(\.markdown)/
 # first = Dir.children(source_dir).first
 # puts first
 # date = first.scan(date_regex)
 # title = first.scan(title_regex)[0][0]
 # puts title
+puts "Copying #{Dir.children(source_dir).count} posts to #{target_dir}"
+Dir.each_child(source_dir) do |child|
+  begin
+    # p "."
+    date = child.scan(/[\d-]*/)
+    title = child.scan(title_regex)[0][0]
 
-Dir.children(source_dir).first do |child|
-  date = child.scan(/[\d-]*/)
-  title = child.scan(title_regex)[0][0]
+    source_path = source_dir + "/" + child
+    target_path = target_dir + "/" + title
 
-  source_path = source_dir + "/" + child
-  target_path = target_dir + "/" + title
+    Dir.mkdir(target_path) unless Dir.exist?(target_path)
+    FileUtils.cp(source_path, target_path + "/index.mdx") unless File.exist?(target_path + "/index.mdx")
+  rescue => e
+    puts e, child
+  end
 
-  Dir.mkdir(target_path) unless Dir.exist?(target_path)
-  FileUtils.cp(source_path, target_path + "/index.mdx")
 end
